Fix what seems to be a broken logic in hildon_vkb_renderer_update_mode
The 'sub keys' loop iterates over the subkeys, however it chagnes state of
the key itself. This doesn't make any sense, esp given that the condition
for the update checks for change in the sub-keys state.

In addition, there is a kind of 'trigger' (at_least_one) so that once
activated, all the keys will be updated no matter if thre is a need or no.

Fix that by implementing what seems to be the correct logic.

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>